Removed contributor
This was originally part of a series of three partner projects for extra credit in a class. But we did it such that my partner did all of project 1, we collaborated on project 2, and I did all of project 3. Now that the class is over, I have removed her name to reflect our contributions.